+++ README.md
@@ -1,3 +1,7 @@
+[<img alt="github" src="https://img.shields.io/badge/github-kakilangit/cetar-37a8e0?style=for-the-badge&labelColor=555555&logo=github" height="20">](https://github.com/kakilangit/cetar)
+[<img alt="crates.io" src="https://img.shields.io/crates/v/cetar.svg?style=for-the-badge&color=ff8b94&logo=rust" height="20">](https://crates.io/crates/cetar)
+[<img alt="docs.rs" src="https://img.shields.io/badge/docs.rs-cetar-bedc9c?style=for-the-badge&labelColor=555555&logo=docs.rs" height="20">](https://docs.rs/cetar)
+
 ![Dall-E generated cetar image](https://raw.githubusercontent.com/kakilangit/cetar/main/static/cetar.png)
 
 # Cetar
