+++ src/lib.rs
@@ -1,4 +1,15 @@
+//! [<img alt="github" src="https://img.shields.io/badge/github-kakilangit/cetar-37a8e0?style=for-the-badge&labelColor=555555&logo=github" height="20">](https://github.com/kakilangit/cetar)
+//! [<img alt="crates.io" src="https://img.shields.io/crates/v/cetar.svg?style=for-the-badge&color=ff8b94&logo=rust" height="20">](https://crates.io/crates/cetar)
+//! [<img alt="docs.rs" src="https://img.shields.io/badge/docs.rs-cetar-bedc9c?style=for-the-badge&labelColor=555555&logo=docs.rs" height="20">](https://docs.rs/cetar)
+//!
+//! 💥 CURL execution timing analyzer
+//!
+//! Cetar is CLI application not a library. But some of the elements can be used as a library.
+//!
+
+/// Color module, contains enum for color and macros to make color.
 pub mod color;
+/// Network module, contains network related functions and structs.
 pub mod network;
+/// Output module, contains display and output handling related functions.
 pub mod output;
-pub mod stat;
